Removing github_id_rsa
The ssh key only came from the container so it's essentially random.  I'm removing the copy from the repo and adding the file to .gitignore so it won't get added again.

Additionally, I think I'll be able to completely do away with the file since I don't really use it.